requirements.txt
fastapi
uvicorn
openai
requests
python-dotenv
python-multipart